refactor(discord): separate adapter responsibilities
